kafka functions

+++ src/models/model.py
@@ -19,4 +19,4 @@ class HappinessPredictions(BASE):
     Region_Asia = Column(Integer, nullable=False)
     Region_Europe = Column(Integer, nullable=False)
     Region_Oceania = Column(Integer, nullable=False)
-    Predited_Score = Column(Integer, nullable=False)
+    Predited_Score = Column(Float, nullable=False)
